+++ Projects/FireAdministrator/Modules/GroupControllerModule/Guard/ViewModels/CodeDetailsViewModel.cs
@@ -52,8 +52,8 @@ namespace GKModule.ViewModels
 			Code.Password = Password;
 		}
 
-		ushort _no;
-		public ushort No
+		int _no;
+		public int No
 		{
 			get { return _no; }
 			set
