+++ Projects/FireAdministrator/Modules/GroupControllerModule/MPTs/ViewModels/MPTDetailsViewModel.cs
@@ -39,8 +39,8 @@ namespace GKModule.ViewModels
 			Description = MPT.Description;
 		}
 
-		ushort _no;
-		public ushort No
+		int _no;
+		public int No
 		{
 			get { return _no; }
 			set
